finnhub_producer/requirements.txt
avro==1.11.3
certifi==2023.7.22
charset-normalizer==3.3.2
finnhub-python==2.4.19
idna==3.4
kafka-python==2.0.2
py4j==0.10.9.7
pyspark==3.5.0
python-dotenv==1.0.0
requests==2.31.0
urllib3==2.1.0
websocket-client==1.6.4
